Refactor & Features: Implemented a controller-service method 'createOrder' to handle the creation of orders based on user input. Added exception handling to catch invalid JSON formats and other exceptions during order creation.

+++ server/src/main/java/com/example/flowersproject/repository/ImageRepository.java
@@ -1,13 +1,11 @@
 package com.example.flowersproject.repository;
 
-import com.example.flowersproject.entity.products.ImageEntity;
+import com.example.flowersproject.entity.product.ImageEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface ImageRepository extends JpaRepository<ImageEntity, Integer> {
+public interface ImageRepository extends JpaRepository<ImageEntity, Long> {
 
     ImageEntity findByImageId(Integer imageId);
